snap/version_update/requirements.txt
argparse==1.4.0
#PyYAML==6.0
ruamel.yaml==0.17.21